services/auth-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.4.0
supabase==2.10.0
httpx==0.27.2
redis==5.2.1
stripe==11.3.0
python-multipart==0.0.22
structlog==24.4.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
slowapi==0.1.9
email-validator==2.2.0
python-dotenv==1.0.1

# Test dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
